Move tensors_about_equal into utils.py in tests
